Let user install even when previous uninstallation failed
If for some reason the uninstaller isn't available to windows anymore,
the installer currently will show an error message and the user has
to workaround it by manually removing the CDDA installer Registry entries.

After this commit, the user will be presented with a
"Try to install anyways?" question, which may fix any issue the player has.
The installer is quite resilient in this case, as it overwrites previously installed
files or adds the missing ones. IMO, it's better than leaving a
non-tech saavy player having to fix errors themselves.