Fix installer not deleting empty directories on uninstall
It seems the Pascal code to delete empty directories wasn't working at all.

The uninstaller was deleting the folders if the installer was the one
creating them, but if for some reason the empty folders existed before installation,
uninstaller wouldn't try to delete them (even if they were empty after uninstalltion).

Anyways, the "uninsalwaysuninstall" flag is exactly what I wanted in the first place.
It deletes empty directories wether installer created them or not, it works recursively
and it doesn't touch any kind of file that the installer didn't install (user files,
log files, etc, which is good to prevent deleting user files by mistake).