Create small commands from ExtractUpdateMessages
* Move ExtractUpdateMessages default parameter values
to three smaller Commands that can be called individually.
This will make it easier to be called from other Commands
* The commands cannot be called with arguments anymore,
but arguably it's not necessary in the context of this project.
As devs will want to run with defaults most of the time.
* If in the future is it necessary to run with any other param,
it's probably worth creating a new command or change the defaults