Change % in translatable HTML string to HTML entity &#37;
* This is to workaround a bug in babel.compile_catalog
which is doing string format check between "100% free software"
and "100% software libre" thinking we are trying to
convert from a "% f" (float) to a "% s" (string) and give us
a validation error. I didn't find a way to disable this behaviour
in babel.compile_catalog and thus the workaround.
Keep in mind that this could be an issue in future translatable strings.