Add The Goat God's Modpack
Again, don't know how I'd put all the mod idents in there, nor who all the authors are. I hope that is acceptable. Also, I'm not entirely certain the link is in a form that the launcher will accept, and there might be a better download link that auto updates.